Revert Material Design 2

+++ brainworkshop.py
@@ -207,6 +207,7 @@ CONFIGFILE_DEFAULT_CONTENTS = """
 # With the default BW sequence generation model, the visual and auditory
 # sequences are more randomized and unpredictable than they are in Jaeggi
 # mode.  The only effect of this option is to set the following options:
+#   OLD_STYLE_SQUARES = True,
 #   SHOW_FEEDBACK = False,
 #   GRIDLINES = False, CROSSHAIRS = True, BLACK_BACKGROUND = True,
 #   WINDOW_FULLSCREEN = True, HIDE_TEXT = True, FIELD_EXPAND = True
@@ -241,6 +242,7 @@ JAEGGI_SCORING = False
 # to emulate the original software used in the study?
 # If this is enabled, the following options will be set:
 #    AUDIO1_SETS = ['letters'],
+#    OLD_STYLE_SQUARES = True,
 #    SHOW_FEEDBACK = False, GRIDLINES = False, CROSSHAIRS = True
 # (note: this option only takes effect if JAEGGI_MODE is set to True)
 # Default: True
@@ -324,6 +326,10 @@ CHANNEL_AUDIO2 = 'right'
 # default?
 # Options: 'color' or 'image'
 MULTI_MODE = 'color'
+
+# Use the flat, single-color squares like in versions prior to 4.1?
+# Also, use sharp corners or rounded corners?
+OLD_STYLE_SQUARES = False
 
 # Start in Manual mode?
 # If this is False, the game will start in standard mode.
@@ -511,6 +517,8 @@ ARITHMETIC_ACCEPTABLE_DECIMALS = ['0.1', '0.2', '0.3', '0.4', '0.5', '0.6',
 
 # Colors for the color n-back task
 # format: (red, green, blue, 255)
+# Note: Changing these colors will have no effect in Dual or
+#   Triple N-Back unless OLD_STYLE_SQUARES is set to True.
 # the _BLK colors are used when BLACK_BACKGROUND is set to True.
 COLOR_1 = (0, 0, 255, 255)
 COLOR_2 = (0, 255, 255, 255)
@@ -744,6 +752,7 @@ if CLINICAL_MODE:
     cfg.SKIP_TITLE_SCREEN                = True
     cfg.USE_MUSIC                        = False
 elif cfg.JAEGGI_INTERFACE_DEFAULT_SCORING:
+    cfg.OLD_STYLE_SQUARES       = True
     cfg.GRIDLINES               = False
     cfg.CROSSHAIRS              = True
     cfg.SHOW_FEEDBACK           = False
@@ -758,6 +767,7 @@ if cfg.JAEGGI_MODE and not cfg.JAEGGI_INTERFACE_DEFAULT_SCORING:
     cfg.JAEGGI_SCORING = True
     if cfg.JAEGGI_FORCE_OPTIONS:
         cfg.AUDIO1_SETS = ['letters']
+        cfg.OLD_STYLE_SQUARES = True
         cfg.GRIDLINES     = False
         cfg.CROSSHAIRS    = True
         cfg.SHOW_FEEDBACK = False
@@ -2213,11 +2223,22 @@ class Visual:
         self.center_y = field.center_y + (field.size // 3)*((position//3+1)%3 - 1)
 
         if self.vis == 0:
-            shape_size = int(field.size / 3 * 0.9)
-            lx = self.center_x - shape_size // 2
-            by = self.center_y - shape_size // 2
-            side = shape_size - 1
-            self.square = pyglet.shapes.Rectangle(lx, by, side, side, color=self.color, batch=batch)
+            if cfg.OLD_STYLE_SQUARES:
+                shape_size = int(field.size / 3 * 0.9)
+                lx = self.center_x - shape_size // 2
+                by = self.center_y - shape_size // 2
+                side = shape_size - 1
+                self.square = pyglet.shapes.Rectangle(lx, by, side, side, color=self.color, batch=batch)
+            else:
+                # use sprite squares
+                shape_size = int(field.size / 3)
+                self.square = self.spr_square[color-1]
+                self.square.opacity = 255
+                self.square.x = self.center_x - field.size // 6
+                self.square.y = self.center_y - field.size // 6
+                self.square.scale = 1.0 * shape_size / self.spr_square_size
+                self.square_size_scaled = self.square.width
+                self.square.batch = batch
 
         elif 'arithmetic' in mode.modalities[mode.mode]: # display a number
             self.label.text = str(number)
@@ -2266,7 +2287,10 @@ class Visual:
                   or (mode.flags[mode.mode]['multi'] > 1 and cfg.MULTI_MODE == 'image'): # hide pictogram
                 self.square.batch = None
             elif self.vis == 0:
-                self.square.delete()
+                if cfg.OLD_STYLE_SQUARES:
+                    self.square.delete()
+                else:
+                    self.square.batch = None
             self.visible = False
 
 # Circles is the 3-strikes indicator in the top left corner of the screen.
